Update django_project.sh
we created so far, we have the colors, which will display red for and error in writing our django app in bash, or green if the script wwas success in creating the file,  first we create our function to have the project name be called tutorial, then the first echo will display what the project name is in our case it is tutorial, we have a echo to create project but if there is not  A list of all double-quoted strings preceded by djangoproject then create the project, and change the dir to djangoproject.  after that we want to create the venv so we do not interact with any current projects going on, then activate the venv, pip install your requirements.txt file, then create a sample of your secret_key and debug mode and allowed_hosts.