Update django_project.sh
what we wrote was we went ahead and copy the End-Of-File to the virtual enviroment on your machine, from there once copied, we should see a green status that the file was success, next we want to create the .gitignore file, that will allow us to ignore certain file that we have listed.  using cat to transfer the data from EOF to .gitignore.